+++ awslimitchecker/services/__init__.py
@@ -47,6 +47,7 @@ from awslimitchecker.services.elasticache import _ElastiCacheService
 from awslimitchecker.services.rds import _RDSService
 from awslimitchecker.services.iam import _IamService
 from awslimitchecker.services.s3 import _S3Service
+from awslimitchecker.services.ses import _SesService
 
 # dynamically generate the service name to class dict
 _services = {}